simplified working tree for main chat component
