+++ screens/main/goal/messageScreen.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import HooksExample from '../chat/components/HooksExample';
+import Chat from '../chat/components/Chat';
 import auth from '@react-native-firebase/auth';
 import { UserContext, ChatContext } from '../chat/contexts'
 import firestore from '@react-native-firebase/firestore';
+import {appStyles} from '../../../assets/styles/styles';
 import {
   SafeAreaView,
   View,
@@ -14,17 +15,16 @@ function MessageScreen({route, navigation}) {
   const {goal} = route.params;
   console.log(goal.params.goal.chatRoomId);
   return (
-    <SafeAreaView style={styles.safe}>
-      <View style={styles.main}>
+    <SafeAreaView style={appStyles.safe}>
+      <View style={appStyles.main}>
         <TouchableOpacity onPress={() => navigation.goBack()}>
           <Text>Go Back</Text>
         </TouchableOpacity>
-        {/* <Text>This is the MessageScreen</Text> */}
       </View>
       {/*need to figure out way to remove tabs in chat*/}
       <ChatContext.Provider value={firestore().collection('ChatRooms').doc(goal.params.goal.chatRoomId).collection('messages')}>
       <UserContext.Provider value={auth().currentUser}>
-        <HooksExample />
+        <Chat />
       </UserContext.Provider>
       </ChatContext.Provider>
     </SafeAreaView>
